+++ package/Runtime/_Shared/SampleScripts.codegen/ArrowShooting.cs
@@ -7,13 +7,28 @@ namespace Needle.Typescript.GeneratedComponents
 {
 	public partial class ArrowShooting : UnityEngine.MonoBehaviour
 	{
-		public UnityEngine.Transform @arrowPrefab;
-		public UnityEngine.AudioSource @sound;
+		public float @power = 1.5f;
+		public float @drawMaxPhysicalDistance = 0.47269f;
+		public float @drawMinPhysicalDistance = 0.1132593f;
+		public UnityEngine.AnimationCurve @drawProgression;
+		public UnityEngine.GameObject @cameraParent;
+		public float @desktopIdleDrawAmount = 0.25f;
+		public UnityEngine.Vector2 @interactionPixelOrigin = new UnityEngine.Vector2(0.5f, 0.65f);
+		public float @interactionPixelTreshold = 85f;
+		public float @drawMaxPixelDistance = 100f;
 		public void awake(){}
+		public void OnEnable(){}
+		public void OnDisable(){}
+		public void mount(){}
+		public void unmount(){}
+		public UnityEngine.GameObject @bowObject;
+		public UnityEngine.Animation @animationComponent;
+		public UnityEngine.Transform @arrowPrefab;
+		public UnityEngine.GameObject @arrowSpawnSpot;
+		public UnityEngine.AudioSource @sound;
+		public float @fakeArrowRespawnDuration = 0.3f;
 		public void onEnterXR(object @_args){}
 		public void onLeaveXR(object @_args){}
-		public UnityEngine.AnimationClip @bowAnimation;
-		public UnityEngine.GameObject @bowObject;
 		public void onBeforeRender(){}
 	}
 }