+++ package/Runtime/_Shared/SampleScripts.codegen/CameraGoal.cs
@@ -7,6 +7,12 @@ namespace Needle.Typescript.GeneratedComponents
 {
 	public partial class CameraGoal : UnityEngine.MonoBehaviour
 	{
+		public bool @createMenuButton = false;
+		public bool @teleportVRPlayer = false;
+		public bool @ignoreYInVR = false;
+		public void awake(){}
+		public void addButton(){}
+		public void removeButton(){}
 		public void use(){}
 	}
 }