+++ package/Runtime/_Shared/SampleScripts~/Bow & Arrow/Arrow.ts
@@ -1,7 +1,5 @@
 import { Behaviour, Collider, Collision, GameObject, Rigidbody, getTempQuaternion, getTempVector } from "@needle-tools/engine";
 
-
-
 export class Arrow extends Behaviour {
 
     destroyTarget: boolean = false;
@@ -14,15 +12,12 @@ export class Arrow extends Behaviour {
         if (this._rigidbody) {
             const col = this.gameObject.getComponentsInChildren(Collider);
             for (const c of col) {
-                const det = GameObject.addNewComponent(c.gameObject, ArrowCollisionDetection);
+                const det = GameObject.addComponent(c.gameObject, ArrowCollisionDetection);
                 det.rigidBody = this._rigidbody;
                 det.arrow = this;
                 det.destroyTarget = this.destroyTarget;
             }
         }
-    }
-
-    start(): void {
         this._startTime = this.context.time.time;
     }
 
@@ -38,7 +33,7 @@ export class Arrow extends Behaviour {
                 const targetRotation = getTempVector().set(vel.x, vel.y, vel.z).normalize();
                 const tempQuat = getTempQuaternion();
                 tempQuat.setFromUnitVectors(getTempVector(0, 0, 1), targetRotation);
-                this.gameObject.quaternion.slerp(tempQuat, this.context.time.deltaTime / .1);
+                this.gameObject.quaternion.copy(tempQuat);
             }
         }
     }
@@ -46,12 +41,15 @@ export class Arrow extends Behaviour {
 
 class ArrowCollisionDetection extends Behaviour {
 
-    arrow!: Arrow;
+    arrow?: Arrow;
     rigidBody!: Rigidbody;
     destroyTarget: boolean = false;
 
     // get stuck when you hit something
     onCollisionEnter(col: Collision) {
+        if (!this.arrow)
+            return;
+
         if (this.destroyTarget) {
             col.gameObject.destroy();
             this.arrow.gameObject.destroy();